Isolate stop hook cap tests
