+++ crates/goose/tests/agent.rs
@@ -496,295 +496,6 @@ mod tests {
     }
 
     #[cfg(test)]
-    mod stop_hook_block_cap_tests {
-        use super::*;
-        use async_trait::async_trait;
-        use goose::agents::{AgentConfig, SessionConfig};
-        use goose::config::permission::PermissionManager;
-        use goose::config::GooseMode;
-        use goose::conversation::message::Message;
-        use goose::model::ModelConfig;
-        use goose::providers::base::{
-            stream_from_single_message, MessageStream, Provider, ProviderUsage, Usage,
-        };
-        use goose::providers::errors::ProviderError;
-        use goose::session::session_manager::SessionType;
-        use goose::session::SessionManager;
-        use rmcp::model::Tool;
-        use serial_test::serial;
-        use std::ffi::OsString;
-        use std::path::{Path, PathBuf};
-        use std::sync::atomic::{AtomicUsize, Ordering};
-        use tempfile::TempDir;
-
-        const ALWAYS_BLOCK_SCRIPT: &str = r#"#!/bin/sh
-echo blocked >> "$PLUGIN_ROOT/hook.log"
-echo "always block" >&2
-exit 2
-"#;
-
-        const ALTERNATE_BLOCK_ALLOW_SCRIPT: &str = r#"#!/bin/sh
-count_file="$PLUGIN_ROOT/count"
-count=0
-if [ -f "$count_file" ]; then
-  count=$(cat "$count_file")
-fi
-count=$((count + 1))
-echo "$count" > "$count_file"
-echo "$count" >> "$PLUGIN_ROOT/hook.log"
-if [ $((count % 2)) -eq 1 ]; then
-  echo "block $count" >&2
-  exit 2
-fi
-exit 0
-"#;
-
-        struct StopHookTestEnv {
-            temp_dir: TempDir,
-            hook_log: PathBuf,
-            prev_cwd: PathBuf,
-            prev_goose_path_root: Option<OsString>,
-            prev_stop_hook_block_cap: Option<OsString>,
-        }
-
-        impl StopHookTestEnv {
-            fn new(stop_hook_block_cap: u32, script: &str) -> Result<Self> {
-                let temp_dir = tempfile::tempdir()?;
-                let project_dir = temp_dir.path().join("project");
-                let plugin_dir = project_dir.join(".agents/plugins/stop-blocker");
-                std::fs::create_dir_all(plugin_dir.join("hooks"))?;
-                std::fs::write(
-                    plugin_dir.join("hooks/hooks.json"),
-                    r#"{
-  "hooks": {
-    "Stop": [
-      {
-        "hooks": [
-          { "type": "command", "command": "sh ${PLUGIN_ROOT}/block.sh" }
-        ]
-      }
-    ]
-  }
-}
-"#,
-                )?;
-                std::fs::write(plugin_dir.join("block.sh"), script)?;
-                std::fs::create_dir_all(temp_dir.path().join("goose-root"))?;
-                std::fs::create_dir_all(temp_dir.path().join("data"))?;
-
-                let prev_cwd = std::env::current_dir()?;
-                let prev_goose_path_root = std::env::var_os("GOOSE_PATH_ROOT");
-                let prev_stop_hook_block_cap = std::env::var_os("GOOSE_STOP_HOOK_BLOCK_CAP");
-
-                unsafe {
-                    std::env::set_var("GOOSE_PATH_ROOT", temp_dir.path().join("goose-root"));
-                    std::env::set_var("GOOSE_STOP_HOOK_BLOCK_CAP", stop_hook_block_cap.to_string());
-                }
-                std::env::set_current_dir(&project_dir)?;
-
-                Ok(Self {
-                    temp_dir,
-                    hook_log: plugin_dir.join("hook.log"),
-                    prev_cwd,
-                    prev_goose_path_root,
-                    prev_stop_hook_block_cap,
-                })
-            }
-
-            fn data_dir(&self) -> PathBuf {
-                self.temp_dir.path().join("data")
-            }
-
-            fn hook_invocations(&self) -> usize {
-                std::fs::read_to_string(&self.hook_log)
-                    .unwrap_or_default()
-                    .lines()
-                    .count()
-            }
-        }
-
-        impl Drop for StopHookTestEnv {
-            fn drop(&mut self) {
-                let _ = std::env::set_current_dir(&self.prev_cwd);
-                restore_env_var("GOOSE_PATH_ROOT", self.prev_goose_path_root.as_ref());
-                restore_env_var(
-                    "GOOSE_STOP_HOOK_BLOCK_CAP",
-                    self.prev_stop_hook_block_cap.as_ref(),
-                );
-            }
-        }
-
-        fn restore_env_var(key: &str, previous: Option<&OsString>) {
-            unsafe {
-                match previous {
-                    Some(value) => std::env::set_var(key, value),
-                    None => std::env::remove_var(key),
-                }
-            }
-        }
-
-        struct CountingTextProvider {
-            call_count: AtomicUsize,
-        }
-
-        impl CountingTextProvider {
-            fn new() -> Self {
-                Self {
-                    call_count: AtomicUsize::new(0),
-                }
-            }
-
-            fn call_count(&self) -> usize {
-                self.call_count.load(Ordering::SeqCst)
-            }
-        }
-
-        #[async_trait]
-        impl Provider for CountingTextProvider {
-            async fn stream(
-                &self,
-                _model_config: &ModelConfig,
-                _session_id: &str,
-                _system_prompt: &str,
-                _messages: &[Message],
-                _tools: &[Tool],
-            ) -> Result<MessageStream, ProviderError> {
-                let call = self.call_count.fetch_add(1, Ordering::SeqCst);
-                let message = Message::assistant().with_text(format!("provider response {call}"));
-                let usage = ProviderUsage::new("mock-model".to_string(), Usage::default());
-                Ok(stream_from_single_message(message, usage))
-            }
-
-            fn get_model_config(&self) -> ModelConfig {
-                ModelConfig::new("mock-model").unwrap()
-            }
-
-            fn get_name(&self) -> &str {
-                "counting-text"
-            }
-        }
-
-        async fn create_agent_and_session(
-            env: &StopHookTestEnv,
-        ) -> Result<(Agent, String, Arc<CountingTextProvider>)> {
-            let session_manager = Arc::new(SessionManager::new(env.data_dir()));
-            let permission_manager = Arc::new(PermissionManager::new(env.data_dir()));
-            let config = AgentConfig::new(
-                session_manager.clone(),
-                permission_manager,
-                None,
-                GooseMode::Auto,
-                true,
-                GoosePlatform::GooseCli,
-            );
-            let agent = Agent::with_config(config);
-            let provider = Arc::new(CountingTextProvider::new());
-            let session = session_manager
-                .create_session(
-                    Path::new(".").to_path_buf(),
-                    "stop-hook-test".to_string(),
-                    SessionType::Hidden,
-                    GooseMode::Auto,
-                )
-                .await?;
-            agent.update_provider(provider.clone(), &session.id).await?;
-            Ok((agent, session.id, provider))
-        }
-
-        async fn run_turn(agent: &Agent, session_id: &str, text: &str) -> Result<Vec<Message>> {
-            let session_config = SessionConfig {
-                id: session_id.to_string(),
-                schedule_id: None,
-                max_turns: Some(10),
-                retry_config: None,
-            };
-            let reply_stream = agent
-                .reply(Message::user().with_text(text), session_config, None)
-                .await?;
-            tokio::pin!(reply_stream);
-
-            let mut messages = Vec::new();
-            while let Some(event) = reply_stream.next().await {
-                match event? {
-                    AgentEvent::Message(message) => messages.push(message),
-                    AgentEvent::McpNotification(_) | AgentEvent::HistoryReplaced(_) => {}
-                }
-            }
-            Ok(messages)
-        }
-
-        fn visible_texts(messages: &[Message]) -> Vec<String> {
-            messages
-                .iter()
-                .map(Message::as_concat_text)
-                .filter(|text| !text.is_empty())
-                .collect()
-        }
-
-        #[tokio::test]
-        #[serial]
-        async fn stop_hook_block_cap_allows_configured_consecutive_blocks_then_overrides(
-        ) -> Result<()> {
-            let env = StopHookTestEnv::new(2, ALWAYS_BLOCK_SCRIPT)?;
-            let (agent, session_id, provider) = create_agent_and_session(&env).await?;
-
-            let messages = run_turn(&agent, &session_id, "hello").await?;
-            let texts = visible_texts(&messages);
-
-            assert_eq!(
-                provider.call_count(),
-                3,
-                "cap=2 should allow two blocked retries, then override on the third block"
-            );
-            assert_eq!(
-                env.hook_invocations(),
-                3,
-                "Stop hook should run for the initial response plus the two honored retries"
-            );
-            assert!(texts.iter().any(|text| text == "provider response 0"));
-            assert!(texts.iter().any(|text| text == "provider response 1"));
-            assert!(texts.iter().any(|text| text == "provider response 2"));
-            assert!(texts.iter().any(|text| {
-                text.contains("more than 2 consecutive times")
-                    && text.contains("GOOSE_STOP_HOOK_BLOCK_CAP")
-            }));
-
-            Ok(())
-        }
-
-        #[tokio::test]
-        #[serial]
-        async fn stop_hook_block_cap_counts_only_consecutive_blocks() -> Result<()> {
-            let env = StopHookTestEnv::new(1, ALTERNATE_BLOCK_ALLOW_SCRIPT)?;
-            let (agent, session_id, provider) = create_agent_and_session(&env).await?;
-
-            let first_turn = run_turn(&agent, &session_id, "first").await?;
-            let second_turn = run_turn(&agent, &session_id, "second").await?;
-            let mut texts = visible_texts(&first_turn);
-            texts.extend(visible_texts(&second_turn));
-
-            assert_eq!(
-                provider.call_count(),
-                4,
-                "each turn should honor one block, retry, then stop when the next Stop hook allows"
-            );
-            assert_eq!(env.hook_invocations(), 4);
-            assert!(texts.iter().any(|text| text == "provider response 0"));
-            assert!(texts.iter().any(|text| text == "provider response 1"));
-            assert!(texts.iter().any(|text| text == "provider response 2"));
-            assert!(texts.iter().any(|text| text == "provider response 3"));
-            assert!(
-                !texts
-                    .iter()
-                    .any(|text| text.contains("overriding and ending turn")),
-                "non-consecutive Stop hook blocks should not trip the cap warning"
-            );
-
-            Ok(())
-        }
-    }
-
-    #[cfg(test)]
     mod tool_pair_summarization_tests {
         use super::*;
         use async_trait::async_trait;
